PR Review - move open_filename to utils
